Fix #19132: Resolve drag-and-drop console error when switching interaction rules (#22382)
Fix drag-and-drop console error when switching interaction rules

The ExpressionChangedAfterItHasBeenCheckedError occurs in Angular
when a value is updated after the change detection cycle has already
been completed. The use of setTimeout to delay the update of
ruleDescriptionFragments introduces the state changes asynchronously,
which could be causing a mismatch in the data. Using ngZone.run()
ensures that updates to ruleDescriptionFragments are executed within
Angular's zone, making Angular aware of the changes and updating its
state without introducing timing issues.